Restructure for weight view, tweak at progress view
