+++ FitnessApp/FitnessApp/Extensions/CurrentWeek.swift
@@ -19,7 +19,7 @@ class CurrentWeek {
     private var calendar = Calendar.current
     
     // Saved Workouts
-    @EnvironmentObject var workouts: WorkoutObject
+    @ObservedObject var workouts = WorkoutObject()
     
     // MARK: - Main Function
     func getCurrentWeek() -> [Day] {
@@ -80,17 +80,17 @@ class CurrentWeek {
         // formats dates to double diget format (03,04,..)
         let dateFormatter = DateFormatter()
         dateFormatter.dateFormat = "dd"
-
+        
         return dateFormatter.string(from: date)
     }
     
     private func getWorkout(atIndex: Int) -> Bool {
-     
+        
+        var done = false
+        
         let lastWorkouts = workouts.savedWorkouts
         
         let currentDate = getDate(atIndex: atIndex)
-        
-        var done = false
         
         for ( _ , value) in lastWorkouts.enumerated() {
             if (value.date.formatted(date: .abbreviated, time: .omitted) == currentDate.formatted(date: .abbreviated, time: .omitted)) {
